+++ app/Reporting/Blocks/Analytics/SiteTrafficBlock.php
@@ -39,6 +39,11 @@ class SiteTrafficBlock extends BlockType
         return 'Analytics';
     }
 
+    public function canBeEmpty(): bool
+    {
+        return true;
+    }
+
     public function requiresCategory(): ?IntegrationCategory
     {
         return IntegrationCategory::Analytics;