+++ app/Reporting/Blocks/Forms/LeadsSummaryBlock.php
@@ -52,6 +52,11 @@ class LeadsSummaryBlock extends BlockType
         return 'Forms & Leads';
     }
 
+    public function canBeEmpty(): bool
+    {
+        return true;
+    }
+
     public function requiresCategory(): ?IntegrationCategory
     {
         return IntegrationCategory::Forms;